requirements_dev.txt
pip==21.3.1
wheel==0.37.1
twine==3.7.1
pytest==6.2.5
paho-mqtt